test_keg: port to generic layer.
