+++ Library/Homebrew/test/test_keg.rb
@@ -304,35 +304,4 @@ class LinkTests < Homebrew::TestCase
     keg.unlink
     keg.uninstall
   end
-
-  def test_mach_o_files_skips_hardlinks
-    a = HOMEBREW_CELLAR/"a/1.0"
-    (a/"lib").mkpath
-    FileUtils.cp dylib_path("i386"), a/"lib/i386.dylib"
-    FileUtils.ln a/"lib/i386.dylib", a/"lib/i386_link.dylib"
-
-    keg = Keg.new(a)
-    keg.link
-
-    assert_equal 1, keg.mach_o_files.size
-  ensure
-    keg.unlink
-    keg.uninstall
-  end
-
-  def test_mach_o_files_isnt_confused_by_symlinks
-    a = HOMEBREW_CELLAR/"a/1.0"
-    (a/"lib").mkpath
-    FileUtils.cp dylib_path("i386"), a/"lib/i386.dylib"
-    FileUtils.ln a/"lib/i386.dylib", a/"lib/i386_link.dylib"
-    FileUtils.ln_s a/"lib/i386.dylib", a/"lib/1.dylib"
-
-    keg = Keg.new(a)
-    keg.link
-
-    assert_equal 1, keg.mach_o_files.size
-  ensure
-    keg.unlink
-    keg.uninstall
-  end
 end
